Refactor delivery report pivot calculations and enhance test assertions
- Updated the `buildPivot` method in `DeliveryReportPivotService` to cast row counts to integers and sums to floats for improved data accuracy.
- Refactored test cases in `DeliveryReportExportTest.php` and `DeliveryReportPivotTest.php` to ensure consistent user authentication and improved clarity in CSV export assertions.

+++ app/Delivery/Services/DeliveryReportPivotService.php
@@ -1492,9 +1492,9 @@ class DeliveryReportPivotService
             foreach ($metrics as $metricIndex => $metricType) {
                 $label = $metricLabels[$metricIndex] ?? ('Metrik '.$metricIndex);
                 if ($metricIndex === 'rows') {
-                    $out[$label] = $row['_count_rows'] ?? 0;
+                    $out[$label] = (int) ($row['_count_rows'] ?? 0);
                 } else {
-                    $out[$label] = $row['_sum_'.$metricIndex] ?? 0;
+                    $out[$label] = (float) ($row['_sum_'.$metricIndex] ?? 0);
                 }
             }
             $result[] = $out;
